requirements.txt
fastapi
uvicorn
pandas
plotly
kaleido
requests
python-dotenv
ta